+++ tests/unit/plugins/modules/test_storagegroup.py
@@ -23,6 +23,8 @@ basic.AnsibleModule = MagicMock()
 
 
 from ansible_collections.dellemc.powermax.plugins.modules.storagegroup import StorageGroup
+from ansible_collections.dellemc.powermax.tests.unit.plugins.module_utils \
+    import mock_storagegroup_api as MockStorageGroupApi
 
 
 class TestStorageGroup():
@@ -31,7 +33,7 @@ class TestStorageGroup():
     sg_args = {'sg_name': 'test_sg', 'service_level': 'None', 'state': 'present',
                'srp': None, 'compression': None, 'volumes': None, 'vol_state': None,
                'child_storage_groups': None, 'child_sg_state': None, 'new_sg_name': None, 'snapshot_policies': None,
-               'snapshot_policy_state': None, 'target_sg_name': None, 'force': None}
+               'snapshot_policy_state': None, 'target_sg_name': None, 'force': None, 'host_io_limit': None}
 
     @pytest.fixture
     def sg_module_mock(self, mocker):
@@ -57,3 +59,39 @@ class TestStorageGroup():
         sg_module_mock.provisioning.get_volume_list = MagicMock(return_value=['vol1', 'vol2'])
         sg_module_mock.perform_module_operation()
         assert "Duplicate volumes found" in sg_module_mock.module.fail_json.call_args[1]['msg']
+
+    def test_move_volumes_between_srfs_protected_sgs(self, sg_module_mock):
+        self.sg_args.update({'state': 'present', 'target_sg_name': 'target_sg', 'volumes': [{'vol_id': '123', 'vol_name': 'test'}],
+                             'vol_state': 'absent-in-group'})
+        sg_module_mock.module.params = self.sg_args
+        sg_module_mock.get_volumes_to_move = MagicMock(return_value=['1', '2', '3'])
+        sg_module_mock.if_srdf_protected = MagicMock(return_value=True)
+        sg_module_mock.perform_module_operation()
+        assert "Specify a force flag to move volumes to or from SRDF protected storage group" \
+            in sg_module_mock.module.fail_json.call_args[1]['msg']
+
+    def test_move_volumes_between_srfs_protected_sgs_with_force(self, sg_module_mock):
+        self.sg_args.update({'state': 'present', 'target_sg_name': 'target_sg', 'volumes': [{'vol_id': '123', 'vol_name': 'test'}],
+                             'vol_state': 'absent-in-group', 'force': True})
+        sg_module_mock.module.params = self.sg_args
+
+        sg_module_mock.get_volumes_to_move = MagicMock(return_value=['1', '2', '3'])
+        sg_module_mock.if_srdf_protected = MagicMock(return_value=True)
+        sg_module_mock.perform_module_operation()
+        assert sg_module_mock.module.exit_json.call_args[1]["modify_sg"]
+        assert sg_module_mock.module.exit_json.call_args[1]["remove_vols_from_sg"]
+
+    def test_set_host_io_limit(self, sg_module_mock):
+        self.sg_args.update({'state': 'present', 'sg_name': 'Test',
+                             'host_io_limit': {'dynamic_distribution': 'Always',
+                                               'host_io_limit_mbps': 100,
+                                               'host_io_limit_iops': 100}})
+        sg_module_mock.module.params = self.sg_args
+        sg_module_mock.provisioning.get_storage_group = MagicMock()
+        sg_module_mock.is_host_io_modification_required = MagicMock()
+        sg_module_mock.validate_host_io_limit_params = MagicMock()
+        sg_module_mock.provisioning.set_host_io_limit_iops_or_mbps = MagicMock(return_value={'host_io_limit_mb_sec': '100',
+                                                                                             'host_io_limit_io_sec': '100',
+                                                                                             'dynamicDistribution': 'Always'})
+        sg_module_mock.perform_module_operation()
+        sg_module_mock.provisioning.set_host_io_limit_iops_or_mbps.assert_called